vjvm: change parameter type of put from ...byte to ...int

+++ vjvm/byte_vector.v
@@ -8,12 +8,12 @@ mut:
 
 pub fn (mut b ByteVec) put_byte(bt int) {
 	b.resize(1)
-	b.put(byte(bt))
+	b.put(bt)
 }
 
 pub fn (mut b ByteVec) put_2byte(bt1 int, bt2 int) {
 	b.resize(2)
-	b.put(byte(bt1), byte(bt2))
+	b.put(bt1, bt2)
 }
 
 fn (mut b ByteVec) resize(incoming_data_len int) {
@@ -22,9 +22,9 @@ fn (mut b ByteVec) resize(incoming_data_len int) {
 	}
 }
 
-fn (mut b ByteVec) put(bts ...byte) {
+fn (mut b ByteVec) put(bts ...int) {
 	for bt in bts {
-		b.data << bt
+		b.data << byte(bt)
 		b.len += 1
 	}
 }
